+++ front/components/Foot.js
@@ -1,13 +1,12 @@
 import React from "react";
-import styles from "../styles/Footer.module.css";
+import styles from "../styles/Foot.module.css";
 import Link from "next/link";
-import useModal from "../utils/useModal";
-import Modal from "./ButtonAsk";
+
+
 
 
 
 export default function Footer() {
-  const { isShowing, toggle } = useModal();
   return (
     <footer className={styles.footer}>
       <div>
@@ -19,8 +18,6 @@ export default function Footer() {
             <Link href="/mentions">Mentions légales</Link>
           </li>
         </ul>
-        <button onClick={toggle}>Poser une question</button>
-        <Modal isShowing={isShowing} hide={toggle} />
       </div>
     </footer>
   );
