+++ front/components/Header.js
@@ -1,55 +1,63 @@
 import {useState} from "react";
 import Link from "next/link";
-import styles from "../styles/Header.module.css";
+import styles from "../styles/Navbar.module.css"
+import * as FaIcons from "react-icons/fa";
+import {HeaderData} from "./HeaderData";
 import useModal from "../utils/useModal";
 import Modal from "./ButtonAsk";
-import * as FaIcons from "react-icons/fa";
-import {HeaderData} from "../components/HeaderData";
 
-export default function Header() {
+
+export default function Navbar() {
   
   const [sidebar, setSidebar] = useState(false);
-  const { isShowing, toggle } = useModal();
   const showSidebar = () => setSidebar(!sidebar);
+  const { isShowing, toggle } = useModal();
   
   return (
-    <header className={styles.header}>
-      <div className={styles.headerWrapper}>
-        <h1>
-          <Link href="/">Les Freelances Lyonnais</Link>
-        </h1>
-        <button onClick={toggle}>Poser une question</button>
-        <Modal isShowing={isShowing} hide={toggle} />
+    <header>
+      <div className={styles.navbar}>
+        
+        <div className={styles.navcontainer}>
+          <div className={styles.leftwrapnav}>
+            <h1>
+              <Link href="/">Les Freelances Lyonnais</Link>
+            </h1>
+          </div>
+
+          <div className={styles.rightwrapnav}>
+            <button  className={styles.navbutton} onClick={toggle}>Poser une question</button>
+            <Modal isShowing={isShowing} hide={toggle} />
+            <FaIcons.FaBars onClick={showSidebar} />
+          </div>
+        </div>
+
       </div>
-      <div className="navbar">
-        {/* <Link as="" href="/" className="menubar">  */}
-          <FaIcons.FaBars onClick={showSidebar} />
-        {/* </Link> */}
-      </div>
-      <nav className={ sidebar ? "navmenu active": "styles.navmenu"  }>
-        <ul className="navmenuitems" onClick={showSidebar}>
-            <li className="navbar-toggle">
-              {/* <Link as="" href="/" className="menubar"> */}
+
+      <nav className={sidebar ? styles["navmenuactive"] : styles["navmenu"] }>
+
+        <ul className={styles["navmenuitems"]} onClick={showSidebar}>
+
+            <li className={styles["navbartoggle"]}>
                 <FaIcons.FaRegTimesCircle/>
-              {/* </Link> */}
             </li>
 
             {HeaderData.map((item, index) => {
               return (
-                <li key={index} className={item.cName}>
+                <li key={index} className={styles["navtextA"]}>
                   
                   <Link as={item.path} href={item.path}>
                     <div>
                     {item.icon}
-                    <span>{item.title}</span>
+                    <span className={styles["myspan"]}>{item.title}</span>
                     </div>
                   </Link>
                 </li>
               );
             })}
 
-        </ul>
+        </ul>        
       </nav>
+
     </header>
   );
 }
