+++ front/pages/test.js
@@ -1,13 +1,14 @@
-import Navbar from "../components/Navbar"
-
-
-
-
-
+import Header from "../components/Header"
+import Main from "../components/Main"
+import Foot from "../components/Foot"
 
 
 export default function Test () {
     return (
-        <Navbar/>
+        <>
+        <Header/>
+        
+        <Foot/>
+        </>
     )
 };
