+++ front/styles/Foot.module.css
@@ -1,0 +1,19 @@
+.footer {
+    width: 100%;
+    height: auto;
+    position: relative;
+    bottom:0;
+    background: black;
+    color: white;
+  }
+  .footer div {
+    align-self: center;
+  }
+  .footer div ul {
+    padding: 15px 0px 15px 75px;
+  }
+  .footer div ul li {
+    cursor: pointer;
+  }
+ 
+
