+++ front/styles/FooterHome.module.css
@@ -1,0 +1,46 @@
+.footer {
+    width: 100%;
+    height: auto;
+    position: relative;
+    bottom:0;
+    background: black;
+    color: white;
+  }
+  .footer div {
+    align-self: center;
+  }
+  .footer div ul {
+    padding: 15px 0px 15px 75px;
+  }
+  .footer div ul li {
+    cursor: pointer;
+  }
+  
+  .footer div button {
+    display: none;
+  }
+  @media screen and (max-width: 768px) {
+    .footer {
+      background-color: transparent;
+      z-index: 1;
+      position: sticky;
+    }
+    .footer div {
+      margin:10px auto;
+      padding:10px 0;
+    }
+    .footer div button {
+      background-color: black;
+      color: white;
+      border-radius: 15px;
+      height: 70px;
+      font-size: 1.25em;
+      width: 300px;
+      display: block;
+      margin: 15px auto;
+      box-shadow: 10px 10px 35px rgba(131, 115, 115, 0.24);
+    }
+    .footer div ul {
+      display: none;
+    }
+  }
